Separate Ascii character math and Utf16 buffer search into dedicated internal utilities
